* move matches models to file * add SITE_ID  env variable * clean up matchdb urls * make MatchInfo camelcase * remove template
