+++ api/matchdb/urls.py
@@ -27,8 +27,6 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('__debug__', include(debug_toolbar.urls)),
     # User
-    path('vsav_info/', include('matches.urls'))
-]
-urlpatterns += [
+    path('vsav_info/', include('matches.urls')),
     path('api-auth/', include('rest_framework.urls')),
 ]
