--- api/matches/templates/test.html
@@ -1,17 +1,0 @@
-<html>
-    <body>
-        <pre>
-            {{type}}
-            {{url}}
-            {{p1_char}}
-            {{p2_char}}
-            {{timestamp}}
-            {{video_title}}
-            {{uploader}}
-            {{date_uploaded}}
-            {{p1_name}}
-            {{p2_name}}
-            {{winner}}
-    </pre>
-    </body>
-</html>
